V4L/DVB: always create a per-file handler data struct
We'll be adding a per-file handler mutex, so we need to always have a
per-fh driver data.

For drivers that need something special, like allocating their on data
that includes inside a v4l2_fh struct, a new method is added to allow
to replace the kernel core struct by a new one.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>