V4L/DVB: videobuf: declare vb_lock as a pointer, instead of an struct
Videobuf-core used to declare the vb_lock mutex, used to protect its data
access during commands that touches on streaming, as an struct built
inside struct videobuf. This requires that all drivers should use the same
schema to protect data, and doesn't allow the drivers or the core to implement
a different locking schema.

By converting it into a pointer, the locking inside videobuf is now optional,
of course, if not declared, the drivers need to take care of the lock by
themselves, or all operations related to videobuf needs to be serialized.

In order to minimize the changes at the drivers, this patch creates a
Temporary global mutex at videobuf, to be replaced by an userspace call at a
next patch. This allows us to do just the core changes and the trivial
changes that needs to happen on all files.

The changes in videobuf were done manually, and the changes inside each driver
were done by this small shell script:

for i in `find drivers/media/ -name *.[ch]`; do echo $i; perl -ne 's/&([\d\w\_\-\>\.]+vb_lock)/$1/; print $_' $i > aa && mv aa $i; done

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>